Renamed resources and packages
