+++ build.gradle.kts
@@ -69,9 +69,9 @@ val generateApiSpec = tasks.register("generateApiSpec",GenerateTask::class){
     setProperty("generatorName", "kotlin-server")
     setProperty("inputSpec",  "$rootDir/spec/swagger.yaml")
     setProperty("outputDir", "$buildDir/generated/api-spec")
-    setProperty("apiPackage", "fi.metatavu.pakkasmarja.services.erp.spec")
-    setProperty("invokerPackage", "fi.metatavu.pakkasmarja.services.erp.invoker")
-    setProperty("modelPackage", "fi.metatavu.pakkasmarja.services.erp.model")
+    setProperty("apiPackage", "${project.group}.api.spec")
+    setProperty("invokerPackage", "${project.group}.api.invoker")
+    setProperty("modelPackage", "${project.group}.api.model")
     this.configOptions.put("library", "jaxrs-spec")
     this.configOptions.put("dateLibrary", "java8")
     this.configOptions.put("interfaceOnly", "true")
@@ -86,7 +86,7 @@ val generateApiClient = tasks.register("generateApiClient",GenerateTask::class){
     setProperty("library", "jvm-okhttp3")
     setProperty("inputSpec",  "$rootDir/spec/swagger.yaml")
     setProperty("outputDir", "$buildDir/generated/api-client")
-    setProperty("packageName", "fi.metatavu.pakkasmarja.services.erp.client")
+    setProperty("packageName", "${project.group}.test.client")
     this.configOptions.put("dateLibrary", "string")
     this.configOptions.put("collectionType", "array")
     this.configOptions.put("serializationLibrary", "jackson")
