+++ src/main/docker/Dockerfile.jvm
@@ -7,18 +7,18 @@
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.jvm -t quarkus/example-api-jvm .
+# docker build -f src/main/docker/Dockerfile.jvm -t quarkus/pakkasmarja-erp-services-jvm .
 #
 # Then run the container using:
 #
-# docker run -i --rm -p 8080:8080 quarkus/example-api-jvm
+# docker run -i --rm -p 8080:8080 quarkus/pakkasmarja-erp-services-jvm
 #
 # If you want to include the debug port into your docker image
 # you will have to expose the debug port (default 5005) like this :  EXPOSE 8080 5050
 #
 # Then run the container using :
 #
-# docker run -i --rm -p 8080:8080 -p 5005:5005 -e JAVA_ENABLE_DEBUG="true" quarkus/example-api-jvm
+# docker run -i --rm -p 8080:8080 -p 5005:5005 -e JAVA_ENABLE_DEBUG="true" quarkus/pakkasmarja-erp-services-jvm
 #
 ###
 FROM registry.access.redhat.com/ubi8/ubi-minimal:8.5
