+++ src/main/docker/Dockerfile.native-distroless
@@ -7,11 +7,11 @@
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.native-distroless -t quarkus/example-api .
+# docker build -f src/main/docker/Dockerfile.native-distroless -t quarkus/pakkasmarja-erp-services .
 #
 # Then run the container using:
 #
-# docker run -i --rm -p 8080:8080 quarkus/example-api
+# docker run -i --rm -p 8080:8080 quarkus/pakkasmarja-erp-services
 #
 ###
 FROM quay.io/quarkus/quarkus-distroless-image:1.0
