+++ src/test/kotlin/fi/metatavu/pakkasmarja/services/erp/test/functional/auth/TestBuilderAuthentication.kt
@@ -1,11 +1,11 @@
-package fi.metatavu.example.api.test.functional.auth
+package fi.metatavu.pakkasmarja.services.erp.test.functional.auth
 
-import fi.metatavu.example.client.infrastructure.ApiClient
-import fi.metatavu.example.api.test.functional.TestBuilder
-import fi.metatavu.example.api.test.functional.impl.ExamplesTestBuilderResource
-import fi.metatavu.example.api.test.functional.settings.ApiTestSettings
 import fi.metatavu.jaxrs.test.functional.builder.auth.AccessTokenProvider
 import fi.metatavu.jaxrs.test.functional.builder.auth.AccessTokenTestBuilderAuthentication
+import fi.metatavu.pakkasmarja.services.erp.test.client.infrastructure.ApiClient
+import fi.metatavu.pakkasmarja.services.erp.test.functional.TestBuilder
+import fi.metatavu.pakkasmarja.services.erp.test.functional.impl.BusinessPartnersTestBuilderResource
+import fi.metatavu.pakkasmarja.services.erp.test.functional.settings.ApiTestSettings
 
 
 /**
@@ -24,7 +24,7 @@ class TestBuilderAuthentication(
 
     private var accessTokenProvider: AccessTokenProvider? = accessTokenProvider
 
-    var examples: ExamplesTestBuilderResource = ExamplesTestBuilderResource(testBuilder, this.accessTokenProvider, createClient())
+    val businessPartners: BusinessPartnersTestBuilderResource = BusinessPartnersTestBuilderResource(testBuilder, this.accessTokenProvider, createClient())
 
     override fun createClient(authProvider: AccessTokenProvider): ApiClient {
         val result = ApiClient(ApiTestSettings.apiBasePath)
